packages/ryoma_lab/pyproject.toml
[build-system]
requires = ["hatchling>=1.4.0"]
build-backend = "hatchling.build"

[project]
name = "ryoma_lab"
version = "0.0.1"
description = "Interactive UI for Ryoma AI Platform"
readme = "README.md"
repository = "https://github.com/project-ryoma/ryoma"
homepage = "https://www.project-ryoma.com"
requires-python = ">=3.10"

dependencies = [
    "alembic>=1.13.1",
    "reflex>=0.5.9, <0.6.0",
    "transformers>=4.40.2",
    "sql-metadata>=2.11.0",
    "feast==0.39.1",
    "amundsen-databuilder>=7.5.0",
    "ibis-framework>=9.0.0",
    "fastapi-users[sqlalchemy,oauth]>=13.0.0",
    "pypdf>=4.3.1",
    "sql-formatter>=0.6.2",
    "torch>=2.4.0",
    "gpt4all>=2.8.2",
    "langchain-huggingface>=0.0.3",
    "langchain-qdrant>=0.1.3",
    "ipython>=8.27.0",
    "jupyter-ai-magics>=2.24.0",
    "langgraph>=0.2.28",
    "nbformat>=5.10.4",
]

[tool.uv.sources]
ryoma_ai = { path = "./packages/ryoma_ai" }

[project.scripts]
ryoma = "ryoma_lab.cli:main"


[project.optional-dependencies]
snowflake = [
    "ibis-framework[snowflake]>=9.0.0",
    "amundsen-databuilder[snowflake]>=7.5.0",
    "feast[snowflake]>=0.39.1"
]
sqlite = [
    "ibis-framework[sqlite]>=9.0.0",
    "feast[sqlite]>=0.39.1",
]
postgres = [
    "psycopg2>=2.9.2",
    "ibis-framework[postgres]>=9.0.0",
    "amundsen-databuilder[postgres]>=7.5.0",
    "feast[postgres]>=0.39.1"
]
mysql = [
    "ibis-framework[mysql]>=9.0.0",
    "amundsen-databuilder[mysql]>=7.5.0",
    "feast[mysql]>=0.39.1"
]
bigquery = [
    "ibis-framework[bigquery]>=9.0.0",
    "amundsen-databuilder[bigquery]>=7.5.0",
    "feast[bigquery]>=0.39.1"
]
pyspark = [
    "ibis-framework[pyspark]>=9.0.0",
    "pyspark>=3.2.0",
    "findspark>=1.4.2",
    "feast[pyspark]>=0.39.1"
]
